Updating to latest changes

+++ src/test/pmt_tests.cpp
@@ -2,18 +2,19 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "arith_uint256.h"
-#include "consensus/merkle.h"
-#include "merkleblock.h"
-#include "serialize.h"
-#include "streams.h"
-#include "test/test_bitcoin.h"
-#include "uint256.h"
-#include "version.h"
+#include <arith_uint256.h>
+#include <consensus/merkle.h>
+#include <merkleblock.h>
+#include <serialize.h>
+#include <streams.h>
+#include <uint256.h>
+#include <version.h>
 
-#include <vector>
+#include <test/test_bitcoin.h>
 
 #include <boost/test/unit_test.hpp>
+
+#include <vector>
 
 class CPartialMerkleTreeTester : public CPartialMerkleTree {
 public:
@@ -86,7 +87,7 @@ BOOST_AUTO_TEST_CASE(pmt_test1) {
 
             // extract merkle root and matched txids from copy
             std::vector<uint256> vMatchTxid2;
-            std::vector<unsigned int> vIndex;
+            std::vector<size_t> vIndex;
             uint256 merkleRoot2 = pmt2.ExtractMatches(vMatchTxid2, vIndex);
 
             // check that it has the same merkle root as the original, and a
@@ -120,7 +121,7 @@ BOOST_AUTO_TEST_CASE(pmt_malleability) {
                                 false, false, false, true,  true,  false};
 
     CPartialMerkleTree tree(vTxid, vMatch);
-    std::vector<unsigned int> vIndex;
+    std::vector<size_t> vIndex;
     BOOST_CHECK(tree.ExtractMatches(vTxid, vIndex).IsNull());
 }
